restructured lots of code
