+++ app/lib/bike/dummy.dart
@@ -1,0 +1,3 @@
+import 'bike.dart';
+
+class DummyBikeConnection implements BikeConnection {}
